Replace pointers-to-members with indexes into methods
There's an msvc code generation issue with constexpr structs that
contain pointers to members, especially in complex types like the
func_or_field variant. Instead, switch to a "dispatch table" style
of referencing chunks of the animation.  This can be further improved
by arrays of function pointers that are just static const and not
constexpr.